requirements.txt
streamlit
requests
fpdf
Pillow
groq
python-dotenv
